+++ CHANGELOG.md
@@ -1,3 +1,49 @@
+# 0.7.4 - 27 Oct 2023
+Improvement:
+- [#24](https://github.com/elysiajs/elysia-swagger/pull/24) - adding schema validity test
+
+Change:
+- [#48](https://github.com/elysiajs/elysia-swagger/pull/48) update Swagger UI to 4.9.0
+- [#36](https://github.com/elysiajs/elysia-swagger/pull/36 ) point to generated .d.ts instead of raw .ts
+
+Bug fix:
+- [#41](https://github.com/elysiajs/elysia-swagger/pull/41) parameters mapping, fix
+- [#43](https://github.com/elysiajs/elysia-swagger/pull/43) typo in default documentation
+
+# 0.7.3 - 26 Sep 2023
+Feature:
+- [#19](https://github.com/elysiajs/elysia-swagger/pull/19) feat: handle nullish response types
+- [#18](https://github.com/elysiajs/elysia-swagger/pull/18) swagger ui options 
+
+
+Improvement:
+- [#23](https://github.com/elysiajs/elysia-swagger/pull/23) Add github action to run bun test
+- remove `removeComment` from tsconfig to show JSDoc
+- add `theme` to customize Swagger CSS link
+- add `autoDarkMode` using poor man Swagger dark mode CSS 😭
+
+Change:
+- Set default swagger version to 5.7.2
+
+Bug fix:
+- [#16](https://github.com/elysiajs/elysia-swagger/pull/16) fix: use global prefix
+
+# 0.7.2 - 21 Sep 2023
+Bug fix:
+- Paths is undefined
+- Models is not showing
+
+# 0.7.1 - 20 Sep 2023
+Bug fix:
+- Add openapi-types as dependencies
+- Fix `any` returned type
+
+# 0.7.0 - 20 Sep 2023
+- Add support for Elysia 0.
+
+# 0.7.0-beta.0 - 18 Sep 2023
+- Add support for Elysia 0.7
+
 # 0.6.2 - 11 Sep 2023
 - Ship lodash.cloneDeep type